Uses server.js as proxy + decoding server to resolve CORS issues. AmtrakAPI.js now serves to provide data structures and functions to use for presentation of api data
