+++ train-tracker/src/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import Amtrak from './AmtrakAPI'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
@@ -9,3 +10,13 @@ root.render(
     <App />
   </React.StrictMode>
 );
+
+
+//example use of API
+const trainData = new Amtrak.TrainData()
+trainData.onUpdated = function () {
+    for(let train of trainData.trains) {
+        console.log(train.toString());
+    }
+}
+trainData.update()
